moving the views and templates
