+++ app/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import Container from './Container';
-import { Section, SectionBody, SectionTitle, SectionPictureBody } from './Section';
-import { MenuOption, Menu, HeaderTitle, Header } from './Header';
+import Container from './views/Container';
+import { Section, SectionBody, SectionTitle, SectionPictureBody } from './views/Section';
+import { MenuOption, Menu, HeaderTitle, Header } from './views/Header';
 
 import './index.css';
 
@@ -34,11 +34,11 @@ ReactDOM.render(
     <Section className="wedding_sectionPicBody">
       <SectionPictureBody urls={
         [
-         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd623e4b0fef6abe1f009/1406916132870/img_0190.jpg?format=750w', 
-         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f00f/1406916134467/img_0346.jpg?format=750w', 
+         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd623e4b0fef6abe1f009/1406916132870/img_0190.jpg?format=750w',
+         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f00f/1406916134467/img_0346.jpg?format=750w',
          'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f013/1406916190056/img_0436-edit.jpg?format=750w',
-         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd623e4b0fef6abe1f009/1406916132870/img_0190.jpg?format=750w', 
-         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f00f/1406916134467/img_0346.jpg?format=750w', 
+         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd623e4b0fef6abe1f009/1406916132870/img_0190.jpg?format=750w',
+         'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f00f/1406916134467/img_0346.jpg?format=750w',
          'https://static1.squarespace.com/static/53d7e714e4b0cb1125a8185a/53dae35ee4b0cfc1a4b33312/53dbd624e4b0fef6abe1f013/1406916190056/img_0436-edit.jpg?format=750w',
         ]
       }/>
