+++ app/views/Section.tsx
